Move doctor helper types
